fix(rollout): share HTTPX response cleanup
