+++ rllm/engine/rollout/tinker_engine.py
@@ -8,6 +8,7 @@ from tinker_cookbook import model_info, renderers
 from tinker_cookbook.renderers import Message
 from typing_extensions import override  # need to use typing_extensions for python < 3.12
 
+from rllm.engine.rollout.httpx_utils import install_httpx_response_cycle_patch
 from rllm.engine.rollout.rollout_engine import ModelOutput, RolloutEngine
 from rllm.engine.rollout.types import ImageProcessor, Processor, TinkerTokenInput, TinkerTokenOutput, TokenInput, Tokenizer, TokenOutput
 from rllm.parser import ChatTemplateParser
@@ -15,6 +16,8 @@ from rllm.tools.tool_base import ToolCall
 from rllm.types import TerminationEvent, TerminationReason
 
 logger = logging.getLogger(__name__)
+
+install_httpx_response_cycle_patch()
 
 """
 Utility functions for Tinker engine. Partly borrowed from
